+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Latest:
    should have used the `strings` package to avoid OS path separator bugs
  - Fixed panic where the Host's contractmanager `AddSectorBatch` allowed for
    writing to a file after the contractmanager had shutdown
+ - Fixed panic where the watchdog would try to write to the contractor's log
+   after the contractor had shutdown
 
 ### v1.4.2.1
 **Key Updates**